+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-- 2024-??-?? 0.4.1
+- 2024-06-14 0.4.2
+    - added Upload button
+    - hardcode initial integration time (400ms) and gain (8dB)
+- 2024-06-14 0.4.1
     - added StatusBar and battery to Scope
 - 2024-06-13 0.4.0
     - fixed colors on Android Go phones